refactor: Computer Class 리팩토링
- Computer Class의 역할을 더 잘 나타내기 위해서 Computer -> ComputerNumbers로 이름 변경
- ComputerNums 필드 추가하고 생성자를 추가하여 객체가 생성될 때 generateNumber() 메서드를 호출하여 숫자를 생성하도록 변경
- generateNumber() 메서드 내부의 상수 값들을 Constants Class에서 가져오도록 변경
